lib: vdso: Add support for time
Add time() vdso support to match up with existing support in the x86's
vdso.  Currently benefitting arm and arm64 which uses the common
vgettimeofday.c implementation.  On arm provides about a ~14 fold
improvement in speed over the straight syscall, and about a ~5 fold
improvement in speed over an alternate library implementation that
relies on the vdso call to gettimeofday to fulfill the request.

We can provide __vdso_time even if we can not provide a speed
enhanced __vdso_gettimeofday.

Bug: 63737556
Bug: 20045882
Signed-off-by: Mark Salyzyn <salyzyn@android.com>
Signed-off-by: wimbiyoashizkia <wimbiyoashizkia@yahoo.com>